requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.8.1=pypi_0
appdirs=1.4.3=py37h28b3542_0
astor=0.8.0=pypi_0
astroid=2.3.2=py37_0
attrs=19.3.0=py_0
black=19.3b0=py_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2019.10.16=0
cachetools=3.1.1=pypi_0
cairo=1.14.12=h8948797_3
certifi=2019.9.11=py37_0
chardet=3.0.4=pypi_0
click=7.0=py37_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
gast=0.2.2=pypi_0
glib=2.56.2=hd408876_0
google-auth=1.6.3=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.1.7=pypi_0
graphite2=1.3.13=h23475e2_0
grpcio=1.24.3=pypi_0
h5py=2.10.0=pypi_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
idna=2.8=pypi_0
intel-openmp=2019.4=243
isort=4.3.21=py37_0
jasper=2.0.14=h07fcdf6_1
jpeg=9b=h024ee3a_2
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
lazy-object-proxy=1.4.2=py37h7b6447c_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=pypi_0
mccabe=0.6.1=py37_1
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.14=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
mss=4.0.3=pypi_0
ncurses=6.1=he6710b0_1
numexpr=2.7.0=pypi_0
numpy=1.17.2=py37haad9e8e_0
numpy-base=1.17.2=py37hde5b4d6_0
oauthlib=3.1.0=pypi_0
openssl=1.1.1d=h7b6447c_3
opt-einsum=3.1.0=pypi_0
pcre=8.43=he6710b0_0
pillow=6.2.1=pypi_0
pip=19.3.1=py37_0
pixman=0.38.0=h7b6447c_0
protobuf=3.10.0=pypi_0
py-opencv=3.4.2=py37hb342d67_1
pyasn1=0.4.7=pypi_0
pyasn1-modules=0.2.7=pypi_0
pylint=2.4.3=py37_0
pynput=1.4.4=pypi_0
python=3.7.4=h265db76_1
python-xlib=0.25=pypi_0
readline=7.0=h7b6447c_5
requests=2.22.0=pypi_0
requests-oauthlib=1.2.0=pypi_0
rsa=4.0=pypi_0
setuptools=41.4.0=py37_0
six=1.12.0=py37_0
sqlite=3.30.1=h7b6447c_0
tables=3.6.0=pypi_0
tensorboard=2.0.1=pypi_0
tensorflow-estimator=2.0.1=pypi_0
tensorflow-gpu=2.0.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.8=hbc83047_0
toml=0.10.0=py37h28b3542_0
urllib3=1.25.6=pypi_0
werkzeug=0.16.0=pypi_0
wheel=0.33.6=py37_0
wrapt=1.11.2=py37h7b6447c_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
